Add support for Python3
Add support for Python3 - part2

Add support for Python3 - part3

Add support for Python3 - part4

Update python3 to work with default encoding different than UTF-8
(for example on RHEL-7 systems)

Support python 3 and also python 2

Don't change unused files.

+++ buildscripts/resmokelib/config.py
@@ -58,7 +58,7 @@ DEFAULTS = {
     "repeat": 1,
     "reportFailureStatus": "fail",
     "reportFile": None,
-    "seed": long(time.time() * 256),  # Taken from random.py code in Python 2.7.
+    "seed": int(time.time() * 256),  # Taken from random.py code in Python 2.7.
     "serviceExecutor": None,
     "shellReadMode": None,
     "shellWriteMode": None,
